Create DDARayTracer.cs
Provides a Cast method which casts a ray between two vectors in tile based coordinates and returns a Vector2 containing the intersection location if an intersection has occurred, using the DDA algorithm. (Very efficient, 5-15ns(0.0015 ms) per ray casted of a length of 100 pixels)

Defines an IRayTraceable interface which allows a caller to pass any type as the Map 2D Array so long as it implements the interface and the Solid boolean property.